requirements.txt
qdrant-client>=1.7.0
fastembed>=0.2.0
torch
transformers
pillow
streamlit
numpy
ultralytics
python-dotenv
openai